Dedicated folder for contract types
