+++ modules/contracts/src/storage/mod.rs
@@ -1,0 +1,30 @@
+// Copyright 2019 by Trinkler Software AG (Switzerland).
+// This file is part of the Katal Chain.
+//
+// Katal Chain is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version <http://www.gnu.org/licenses/>.
+//
+// Katal Chain is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+use super::*;
+
+mod events;
+mod states;
+mod terms;
+
+pub use events::*;
+pub use states::*;
+pub use terms::*;
+
+// This struct contains all the information that defines a contract state.
+#[derive(Clone, Decode, Debug, Encode, Default, PartialEq)]
+pub struct Contract {
+    pub terms: Terms,
+    pub states: States,
+    pub schedule: Vec<Event>,
+}
