+++ modules/contracts/src/types/mod.rs
@@ -13,12 +13,8 @@
 
 use super::*;
 
-mod deploy_ann;
-mod deploy_pam;
-mod progress_ann;
-mod progress_pam;
+mod ann;
+mod pam;
 
-pub use deploy_ann::*;
-pub use deploy_pam::*;
-pub use progress_ann::*;
-pub use progress_pam::*;
+pub use ann::*;
+pub use pam::*;
